Split in_multnomah_county_or into separate file
Per reviewer comment, moved the in_multnomah_county_or variable
to its own file following PolicyEngine convention of one variable
per file.
